Add tiles
This commit, the first on branch worker-tiles, attempts to add tiles to the version with a single worker. There are various issues with this as yet, so, for now, I've set it to the degenerate case: just one tile. In general, the tiles only works for a single row (multiple rows duplicate the whole shape). The messages coming in asynchronously either result in poor control (overshooting), if all are processed since the last reset, or clunkiness if only the latest batch of tile calculations are rendered. I also introduced a perimeter check in the Rust function, calculate_mandelbrot. With more than one tile, this seems to generate false positives, coloring teh whole tile black sometimes even when it shouldn't be, or perhaps the tile was correctly colored all black but placed wrongly over its neighbour?